+++ src/main/java/study/notes/API 개발 고급 정리.md
@@ -16,7 +16,8 @@
 ⭐권장 순서
 1. 엔티티 조회 방식으로 우선 접근
    1. 페치 조인으로 쿼리 수 최적화
-   2. 컬렉션 최적화 :  페이징 필요하면 batchSize 설정으로 최적화하고, 페이징 필요 없다면 페치 조인 사용해도됨.
+   2. 컬렉션 최적화 :  페이징 필요하면 지연로딩으로 유지하고 n+1 문제해결을 위해 batchSize 설정으로 최적화한다.
+   페이징 필요 없다면 페치 조인 사용해도됨.
    
 2. 엔티티 조회 방식으로 해결 안되면 DTO 직접 조회 방식 사용
 3. DTO 조회 방식으로 해결이 안되면 NativeSQL OR 스프링 jdbc Templet활용
